fix: fix crash on 32-bit devices caused by Double -> Int conversion

+++ RInAppMessaging/Classes/Events/AppStartEvent.swift
@@ -15,7 +15,7 @@
                    name: Constants.Event.appStart)
     }
 
-    init(timestamp: Int) {
+    init(timestamp: Int64) {
         super.init(type: EventType.appStart,
                    name: Constants.Event.appStart,
                    timestamp: timestamp)
