+++ RInAppMessaging/Classes/Extensions/Date+IAM.swift
@@ -1,6 +1,6 @@
 /// Extension to Date class to provide computed properties required by InAppMessaging.
 internal extension Date {
-    var millisecondsSince1970: Int {
-        return Int((self.timeIntervalSince1970 * 1000.0).rounded())
+    var millisecondsSince1970: Int64 {
+        return Int64((self.timeIntervalSince1970 * 1000.0).rounded())
     }
 }
